SPARK-130 Moved rehash command to administration commands module  - also made administration commands module a thing.
